refactor(v2.6.0): extract modular foundation package
- Create foundation/ package with utils, config, and education modules
- Extract constants and helper functions to foundation/utils.py
- Extract config loading and argument parsing to foundation/config.py
- Extract educational content to foundation/education.py
- Create foundation/templates/ stub for future template extraction
- Update setup_foundation_lite.py to import from foundation package
- Reduce main script from ~2177 lines to ~1564 lines (-613 lines)
- All 19 smoke tests pass

Closes #27, Closes #28, Closes #29
