+++ foundation/templates/__init__.py
@@ -1,0 +1,20 @@
+# foundation/templates - Template generation subpackage
+
+"""
+Templates subpackage for Project Foundation Template.
+
+Modules:
+    core: README, CONTRIBUTING, LICENSE, .gitignore
+    governance: CoC, SECURITY, CHANGELOG
+    github: Issue templates, PR template, workflows
+    advanced: ADR, CI, secrets detection
+
+Note: Template extraction will be completed incrementally.
+Currently, templates are still generated by the main generator class.
+"""
+
+# Template modules will be imported here as they are extracted
+# from foundation.templates.core import ...
+# from foundation.templates.governance import ...
+# from foundation.templates.github import ...
+# from foundation.templates.advanced import ...
