+++ setup_foundation_lite.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 """
 Project Foundation Generator - Ethical Lite Edition
-Version: 2.1.0-lite
+Version: 2.6.0-lite
 Released: 2025-09-24
 
 This tool generates educational project foundations with built-in governance principles.
@@ -27,426 +27,32 @@ limitations under the License.
 Repository: https://github.com/malcolmhoward/project-foundation-template
 """
 
-import argparse
 import json
-import os
 import sys
-import hashlib
 import time
-from datetime import date, datetime, timedelta
+from datetime import date, datetime
 from pathlib import Path
-from typing import Dict, Set, Any
+
+# v2.6.0: Import from modular foundation package
+from foundation import (
+    # Utils
+    SCRIPT_VERSION,
+    OFFICIAL_REPO,
+    EXPIRATION_DATE,
+    configure_windows_utf8,
+    # Config
+    load_config_file,
+    merge_config_with_args,
+    parse_arguments,
+    validate_arguments,
+    # Education
+    LITE_PRINCIPLES,
+    EDUCATION_CONTENT,
+    ETHICAL_USE_AGREEMENT,
+)
 
 # Configure stdout for UTF-8 on Windows (handles Unicode characters in output)
-if sys.platform == 'win32':
-    try:
-        sys.stdout.reconfigure(encoding='utf-8')
-        sys.stderr.reconfigure(encoding='utf-8')
-    except AttributeError:
-        # Python < 3.7 fallback
-        import io
-        sys.stdout = io.TextIOWrapper(sys.stdout.buffer, encoding='utf-8')
-        sys.stderr = io.TextIOWrapper(sys.stderr.buffer, encoding='utf-8')
-
-# Version and expiration
-SCRIPT_VERSION = "2.5.0-lite"
-EXPIRATION_DATE = date(2026, 3, 1)
-OFFICIAL_REPO = "https://github.com/malcolmhoward/project-foundation-template"
-
-# Default config file names (searched in order)
-CONFIG_FILES = [".foundationrc", ".foundationrc.json", "foundationrc.json"]
-
-# v2.4.0: Secrets patterns for detection
-SECRETS_PATTERNS = [
-    r"(api[_-]?key|apikey)['\"]?\s*[:=]\s*['\"]?[A-Za-z0-9+/]{20,}",
-    r"(secret|token|password|passwd|pwd)['\"]?\s*[:=]\s*['\"]?[A-Za-z0-9+/]{8,}",
-    r"(aws[_-]?access[_-]?key[_-]?id|aws[_-]?secret[_-]?access[_-]?key)",
-    r"-----BEGIN (RSA |EC |OPENSSH |DSA |SSH2 |)PRIVATE KEY-----",
-    r"(ghp|gho|ghs|ghu)_[A-Za-z0-9_]{36,}",  # GitHub tokens
-    r"sk-[A-Za-z0-9]{48}",  # OpenAI keys
-    r"(mongodb(\+srv)?|postgres(ql)?|mysql|redis)://[^\\s]+",  # Database URLs
-    r"eyJ[A-Za-z0-9+/]*\\.eyJ[A-Za-z0-9+/]*\\.[A-Za-z0-9+/_-]*"  # JWT tokens
-]
-
-
-def load_config_file(config_path: str = None) -> Dict[str, Any]:
-    """
-    Load configuration from a .foundationrc file.
-
-    Searches for config files in order:
-    1. Explicit path if provided via --config
-    2. .foundationrc in current directory
-    3. .foundationrc.json in current directory
-    4. foundationrc.json in current directory
-
-    Returns empty dict if no config found (not an error).
-    """
-    config = {}
-
-    if config_path:
-        # Explicit config path provided
-        path = Path(config_path)
-        if not path.exists():
-            print(f"⚠️  Config file not found: {config_path}")
-            print("   Continuing with command-line arguments only.\n")
-            return {}
-        try:
-            with open(path, 'r', encoding='utf-8') as f:
-                config = json.load(f)
-            print(f"📋 Loaded config from: {path}\n")
-            return config
-        except json.JSONDecodeError as e:
-            print(f"❌ Error parsing config file: {e}")
-            print("   Config file must be valid JSON.\n")
-            return {}
-
-    # Search for default config files
-    for filename in CONFIG_FILES:
-        path = Path(filename)
-        if path.exists():
-            try:
-                with open(path, 'r', encoding='utf-8') as f:
-                    config = json.load(f)
-                print(f"📋 Loaded config from: {path}\n")
-                return config
-            except json.JSONDecodeError:
-                continue  # Try next file
-
-    return config
-
-
-def merge_config_with_args(args, config: Dict[str, Any]):
-    """
-    Merge config file settings with command-line arguments.
-    Command-line arguments take precedence over config file.
-    """
-    # Map config keys to argument names
-    config_mapping = {
-        "project_name": "project_name",
-        "project-name": "project_name",
-        "projectName": "project_name",
-        "author_name": "author_name",
-        "author-name": "author_name",
-        "authorName": "author_name",
-        "license": "license",
-        "output_dir": "output_dir",
-        "output-dir": "output_dir",
-        "outputDir": "output_dir",
-        "include_coc": "include_coc",
-        "include-coc": "include_coc",
-        "includeCoc": "include_coc",
-        "include_security": "include_security",
-        "include-security": "include_security",
-        "includeSecurity": "include_security",
-        # v2.3.0: Community governance templates
-        "include_github_templates": "include_github_templates",
-        "include-github-templates": "include_github_templates",
-        "includeGithubTemplates": "include_github_templates",
-        "include_changelog": "include_changelog",
-        "include-changelog": "include_changelog",
-        "includeChangelog": "include_changelog",
-        "include_all": "include_all",
-        "include-all": "include_all",
-        "includeAll": "include_all",
-        # v2.4.0: Security features
-        "include_enhanced_security": "include_enhanced_security",
-        "include-enhanced-security": "include_enhanced_security",
-        "includeEnhancedSecurity": "include_enhanced_security",
-        "include_secrets_detection": "include_secrets_detection",
-        "include-secrets-detection": "include_secrets_detection",
-        "includeSecretsDetection": "include_secrets_detection",
-        # v2.5.0: Advanced governance
-        "include_adr": "include_adr",
-        "include-adr": "include_adr",
-        "includeAdr": "include_adr",
-        "include_ci": "include_ci",
-        "include-ci": "include_ci",
-        "includeCi": "include_ci",
-        # v2.2.0: Non-interactive mode
-        "verbose": "verbose",
-        "non_interactive": "non_interactive",
-        "non-interactive": "non_interactive",
-        "nonInteractive": "non_interactive",
-        "accept_terms": "accept_terms",
-        "accept-terms": "accept_terms",
-        "acceptTerms": "accept_terms",
-    }
-
-    for config_key, arg_name in config_mapping.items():
-        if config_key in config:
-            # Only apply config value if arg wasn't explicitly set
-            current_value = getattr(args, arg_name, None)
-            default_values = {
-                "project_name": None,
-                "author_name": None,
-                "license": "mit",
-                "output_dir": ".",
-                "include_coc": False,
-                "include_security": False,
-                # v2.3.0
-                "include_github_templates": False,
-                "include_changelog": False,
-                "include_all": False,
-                # v2.4.0
-                "include_enhanced_security": False,
-                "include_secrets_detection": False,
-                # v2.5.0
-                "include_adr": False,
-                "include_ci": False,
-                # v2.2.0
-                "verbose": False,
-                "non_interactive": False,
-                "accept_terms": False,
-            }
-            if current_value == default_values.get(arg_name):
-                setattr(args, arg_name, config[config_key])
-
-    return args
-
-
-# Ethical use agreement that must be acknowledged
-ETHICAL_USE_AGREEMENT = """
-╔══════════════════════════════════════════════════════════════════════════════╗
-║                     ETHICAL USE AGREEMENT - PLEASE READ                       ║
-╠══════════════════════════════════════════════════════════════════════════════╣
-║                                                                                ║
-║  By using this tool, you acknowledge and agree that:                          ║
-║                                                                                ║
-║  1. This generates TEMPLATES ONLY - not complete solutions                    ║
-║  2. Templates MUST be customized for your specific needs                      ║
-║  3. This does NOT provide legal or compliance advice                          ║
-║  4. This does NOT make your project automatically secure                      ║
-║  5. This does NOT constitute professional security consultation               ║
-║  6. You will implement these practices, not just generate documents           ║
-║  7. You will not misrepresent templates as professional compliance            ║
-║  8. You accept full responsibility for how you use these templates            ║
-║                                                                                ║
-║  Misrepresenting these templates as actual compliance or security             ║
-║  certification may constitute fraud and could result in:                      ║
-║  - Legal penalties                                                            ║
-║  - Data breaches                                                              ║
-║  - Loss of user trust                                                         ║
-║  - Personal liability                                                         ║
-║                                                                                ║
-║  This tool exists to democratize best practices and education.                ║
-║  Please use it responsibly and ethically.                                     ║
-║                                                                                ║
-╚══════════════════════════════════════════════════════════════════════════════╝
-"""
-
-# Core governance principles for lite version
-LITE_PRINCIPLES = {
-    "readme": {
-        "name": "README Documentation",
-        "why": "First impression and guide for users and contributors",
-        "what": "Clear project overview with purpose, setup, and usage",
-        "risk": "Without it, projects are unapproachable and unused"
-    },
-    "contributing": {
-        "name": "Contribution Guidelines", 
-        "why": "Enables community collaboration and maintains quality",
-        "what": "Clear process for contributions and expectations",
-        "risk": "Without it, contributions are chaotic or non-existent"
-    },
-    "license": {
-        "name": "License Selection",
-        "why": "Defines legal terms for use and contribution",
-        "what": "Clear intellectual property terms",
-        "risk": "Without it, legal ambiguity prevents adoption"
-    },
-    "code-of-conduct": {
-        "name": "Code of Conduct",
-        "why": "Creates safe, inclusive environment for all",
-        "what": "Behavioral expectations and enforcement",
-        "risk": "Without it, toxic behavior drives away contributors"
-    },
-    "security": {
-        "name": "Security Basics",
-        "why": "Vulnerabilities can harm users and reputation",
-        "what": "Basic security practices and reporting process",
-        "risk": "Without it, projects become attack vectors"
-    },
-    # v2.4.0: Enhanced security features
-    "enhanced-security": {
-        "name": "Enhanced Security Policy",
-        "why": "Comprehensive security documentation builds trust and enables responsible disclosure",
-        "what": "Detailed SECURITY.md with vulnerability reporting, response timelines, and best practices",
-        "risk": "Without clear security processes, vulnerabilities go unreported or are disclosed publicly"
-    },
-    "secrets-detection": {
-        "name": "Secrets Detection",
-        "why": "Accidentally committed secrets are a leading cause of breaches",
-        "what": "Pre-commit hooks that scan for API keys, passwords, and tokens before they enter history",
-        "risk": "Once secrets are in git history, they're nearly impossible to fully remove"
-    },
-    # v2.5.0: Advanced governance
-    "adr": {
-        "name": "Architecture Decision Records",
-        "why": "Teams forget why decisions were made, leading to repeated debates or reversed progress",
-        "what": "Lightweight documents capturing context, decision, and consequences of architectural choices",
-        "risk": "Without ADRs, institutional knowledge leaves when team members do"
-    },
-    "ci-workflow": {
-        "name": "CI/CD Workflow",
-        "why": "Manual testing is error-prone and inconsistent",
-        "what": "GitHub Actions workflow for automated testing, linting, and validation",
-        "risk": "Without CI, bugs slip through and code quality degrades over time"
-    },
-    # v2.3.0: Community governance templates
-    "issue-templates": {
-        "name": "Issue Templates",
-        "why": "Structured bug reports and feature requests save time",
-        "what": "Templates that guide users to provide needed information",
-        "risk": "Without them, issues lack context and take longer to resolve"
-    },
-    "pr-template": {
-        "name": "Pull Request Template",
-        "why": "Consistent PR descriptions improve review quality",
-        "what": "Checklist ensuring code quality and documentation",
-        "risk": "Without it, PRs are inconsistent and harder to review"
-    },
-    "changelog": {
-        "name": "Changelog",
-        "why": "Users need to know what changed between versions",
-        "what": "Human-readable history of notable changes",
-        "risk": "Without it, users can't assess upgrade impact"
-    }
-}
-
-# Educational content for each principle
-EDUCATION_CONTENT = {
-    "readme": """
-    📚 LEARNING: The README is your project's front door.
-    
-    Studies show that projects with clear READMEs get 5x more contributions.
-    A good README answers: What? Why? How? Who? When?
-    
-    Without a README, even excellent code goes unused because people
-    don't understand what it does or how to use it.
-    """,
-    
-    "contributing": """
-    📚 LEARNING: Contribution guidelines are your quality gateway.
-    
-    Clear contribution guidelines reduce review time by 50% and increase
-    contribution quality. They set expectations and prevent frustration.
-    
-    Without them, you'll spend more time rejecting or fixing contributions
-    than it would take to write the guidelines.
-    """,
-    
-    "license": """
-    📚 LEARNING: No license means "all rights reserved" by default.
-    
-    Without a license, others legally cannot use, modify, or contribute
-    to your project. This is the opposite of what most people intend.
-    
-    Choose MIT for maximum freedom, GPL for copyleft, or Apache for
-    patent protection. When in doubt, use MIT.
-    """,
-    
-    "code-of-conduct": """
-    📚 LEARNING: Codes of Conduct increase diversity by 40%.
-    
-    Research shows that projects with CoCs have more diverse contributors
-    and fewer toxic interactions. They're not about policing, but about
-    setting positive expectations.
-    
-    Without one, you implicitly tolerate behavior that drives people away.
-    """,
-    
-    "security": """
-    📚 LEARNING: 83% of projects have at least one vulnerability.
-
-    Security isn't optional - it's a responsibility to your users.
-    Basic practices like dependency scanning and security policies
-    can prevent most common vulnerabilities.
-
-    Without security practices, you're one CVE away from headlines.
-    """,
-
-    # v2.4.0: Enhanced security features
-    "enhanced-security": """
-    📚 LEARNING: A clear security policy enables responsible vulnerability disclosure.
-
-    Security researchers need to know how to report issues safely. Without a
-    SECURITY.md, they may disclose publicly, report to the wrong channel, or
-    not report at all. GitHub recommends security policies for all public repos.
-    (Reference: https://docs.github.com/en/code-security/getting-started/adding-a-security-policy-to-your-repository)
-
-    Key elements: Supported versions, reporting process, response timeline, and
-    security best practices specific to your project.
-    """,
-
-    "secrets-detection": """
-    📚 LEARNING: Exposed credentials are consistently among the top causes of breaches.
-
-    According to the Verizon Data Breach Investigations Report, stolen/compromised
-    credentials are involved in a significant portion of breaches each year.
-    (Reference: https://www.verizon.com/business/resources/reports/dbir/)
-
-    Once a secret is committed to git, it's in the history forever - even if you
-    delete it from the current version. Pre-commit hooks catch secrets BEFORE
-    they become permanent security risks.
-
-    Common patterns detected: API keys, passwords, tokens, private keys, and
-    database connection strings.
-    """,
-
-    # v2.5.0: Advanced governance
-    "adr": """
-    📚 LEARNING: Architecture Decision Records preserve institutional knowledge.
-
-    Michael Nygard introduced ADRs in 2011 as a way to capture the 'why' behind
-    architectural decisions. Without them, teams often reverse good decisions
-    because they don't understand the original context.
-    (Reference: https://cognitect.com/blog/2011/11/15/documenting-architecture-decisions)
-
-    ADRs are lightweight - just markdown files with context, decision, and
-    consequences. They're versioned with the code they describe.
-    """,
-
-    "ci-workflow": """
-    📚 LEARNING: CI/CD catches issues before they reach production.
-
-    Continuous Integration ensures that every change is automatically tested.
-    GitHub Actions provides free CI for public repositories and integrates
-    directly with pull requests.
-    (Reference: https://docs.github.com/en/actions/automating-builds-and-tests)
-
-    A basic CI workflow runs tests, linting, and builds on every PR, giving
-    reviewers confidence that the code works.
-    """,
-
-    # v2.3.0: Community governance templates
-    "issue-templates": """
-    📚 LEARNING: Structured issue templates reduce resolution time by 30%.
-
-    When users report bugs without environment info, or request features
-    without context, maintainers waste time asking follow-up questions.
-
-    Good templates guide users to provide what you need upfront.
-    """,
-
-    "pr-template": """
-    📚 LEARNING: Projects with PR templates have 50% fewer back-and-forth reviews.
-
-    A PR checklist reminds contributors to run tests, update docs, and
-    describe their changes. It sets quality expectations before review.
-
-    Without it, reviewers must manually check for common oversights.
-    """,
-
-    "changelog": """
-    📚 LEARNING: 70% of users check changelogs before upgrading.
-
-    A changelog is your project's narrative - it tells users what to
-    expect from each version. Keep a Changelog format (keepachangelog.com)
-    is the de facto standard.
-
-    Without it, users fear breaking changes and delay upgrades.
-    """
-}
+configure_windows_utf8()
 
 class EthicalFoundationGenerator:
     """Main generator class with ethical safeguards and education."""
@@ -2161,204 +1767,6 @@ Together, we can build better, safer software.
 
 Learn more: {OFFICIAL_REPO}
 """)
-
-def parse_arguments():
-    """Parse command line arguments."""
-    parser = argparse.ArgumentParser(
-        description="Generate ethical project foundations with education",
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-        epilog="""
-This tool generates TEMPLATE documentation for project governance.
-Templates must be customized for your specific needs.
-
-Examples:
-  Interactive mode:
-    %(prog)s --project-name MyProject --author-name "Jane Doe"
-
-  Non-interactive mode (for CI/scripts):
-    %(prog)s --non-interactive --accept-terms --project-name MyProject --author-name "Jane Doe"
-
-  Using config file:
-    %(prog)s --config .foundationrc
-
-Config file format (.foundationrc):
-  {
-    "project_name": "MyProject",
-    "author_name": "Jane Doe",
-    "license": "mit",
-    "include_coc": true,
-    "include_security": true
-  }
-        """
-    )
-
-    # Core options
-    parser.add_argument(
-        "--project-name",
-        dest="project_name",
-        help="Name of your project (required unless in config)"
-    )
-
-    parser.add_argument(
-        "--author-name",
-        dest="author_name",
-        help="Your name or organization (required unless in config)"
-    )
-
-    parser.add_argument(
-        "--license",
-        choices=["mit", "apache", "gpl"],
-        default="mit",
-        help="License type (default: mit)"
-    )
-
-    parser.add_argument(
-        "--output-dir",
-        dest="output_dir",
-        default=".",
-        help="Output directory (default: current directory)"
-    )
-
-    # Optional components
-    parser.add_argument(
-        "--include-coc",
-        dest="include_coc",
-        action="store_true",
-        help="Include Code of Conduct template"
-    )
-
-    parser.add_argument(
-        "--include-security",
-        dest="include_security",
-        action="store_true",
-        help="Include Security Policy template"
-    )
-
-    # v2.3.0: Community governance templates
-    parser.add_argument(
-        "--include-github-templates",
-        dest="include_github_templates",
-        action="store_true",
-        help="Include GitHub issue and PR templates (.github/ folder)"
-    )
-
-    parser.add_argument(
-        "--include-changelog",
-        dest="include_changelog",
-        action="store_true",
-        help="Include CHANGELOG.md template"
-    )
-
-    parser.add_argument(
-        "--all",
-        dest="include_all",
-        action="store_true",
-        help="Include all optional templates (CoC, Security, GitHub, Changelog, Enhanced Security, Secrets Detection, ADR, CI)"
-    )
-
-    # v2.4.0: Security features
-    parser.add_argument(
-        "--include-enhanced-security",
-        dest="include_enhanced_security",
-        action="store_true",
-        help="Include comprehensive SECURITY.md with vulnerability reporting process"
-    )
-
-    parser.add_argument(
-        "--include-secrets-detection",
-        dest="include_secrets_detection",
-        action="store_true",
-        help="Include pre-commit config with secrets detection hooks"
-    )
-
-    # v2.5.0: Advanced governance
-    parser.add_argument(
-        "--include-adr",
-        dest="include_adr",
-        action="store_true",
-        help="Include Architecture Decision Record templates"
-    )
-
-    parser.add_argument(
-        "--include-ci",
-        dest="include_ci",
-        action="store_true",
-        help="Include GitHub Actions CI workflow"
-    )
-
-    # v2.2.0: Non-interactive mode
-    parser.add_argument(
-        "--non-interactive",
-        dest="non_interactive",
-        action="store_true",
-        help="Run without prompts (for CI/scripts). Requires --accept-terms."
-    )
-
-    parser.add_argument(
-        "--accept-terms",
-        dest="accept_terms",
-        action="store_true",
-        help="Accept ethical use agreement (required for --non-interactive)"
-    )
-
-    # v2.2.0: Config file support
-    parser.add_argument(
-        "--config",
-        dest="config_path",
-        help="Path to config file (default: .foundationrc in current directory)"
-    )
-
-    # v2.2.0: JSON export
-    parser.add_argument(
-        "--export-json",
-        dest="export_json",
-        action="store_true",
-        help="Output results as JSON (useful for tooling integration)"
-    )
-
-    # Output control
-    parser.add_argument(
-        "--verbose",
-        action="store_true",
-        help="Show detailed output"
-    )
-
-    parser.add_argument(
-        "--quiet",
-        action="store_true",
-        help="Minimal output (implies --non-interactive behavior for output only)"
-    )
-
-    parser.add_argument(
-        "--version",
-        action="version",
-        version=f"%(prog)s {SCRIPT_VERSION}"
-    )
-
-    return parser.parse_args()
-
-def validate_arguments(args) -> tuple:
-    """
-    Validate that required arguments are provided.
-    Returns (is_valid, error_message).
-    """
-    errors = []
-
-    if not args.project_name:
-        errors.append("--project-name is required (or set 'project_name' in config file)")
-
-    if not args.author_name:
-        errors.append("--author-name is required (or set 'author_name' in config file)")
-
-    if args.non_interactive and not args.accept_terms:
-        errors.append("--accept-terms is required when using --non-interactive mode")
-        errors.append("  This ensures you've read and understood the ethical use agreement.")
-        errors.append("  Review the agreement by running without --non-interactive first.")
-
-    if errors:
-        return False, "\n".join(errors)
-
-    return True, ""
 
 
 def main():
